Time: 640 ms (18.57%), Space: 174.2 MB (16.83%) - LeetHub

+++ 0786-k-th-smallest-prime-fraction/0786-k-th-smallest-prime-fraction.cpp
@@ -6,7 +6,7 @@ public:
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < n; j++) {
                 if (i != j) {
-                    pq.push({arr[i] / arr[j], {arr[i], arr[j]}});
+                    pq.push({(float)arr[i] / (float)arr[j], {arr[i], arr[j]}});
                 }
             }
         }
